requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
bcrypt<4.1
python-multipart
